+++ packages/router-core/src/load-client.ts
@@ -2,14 +2,16 @@
 // can rewrite relative imports for both ESM and CJS.
 import { isNotFound } from './not-found'
 import { isRedirect } from './redirect'
-import { loadRouteChunk } from './route-chunks'
-import { _getRenderedMatches } from './rendered-matches'
 import {
   _getUserHistoryState,
   getLocationChangeInfo,
   runRouteLifecycle,
 } from './router'
 import { deepEqual } from './utils'
+import { hydrateSsrMatchId } from './ssr/ssr-match-id'
+import type { GLOBAL_SEROVAL, GLOBAL_TSR } from './ssr/constants'
+import type { AnySerializationAdapter } from './ssr/serializer/transformer'
+import type { TsrSsrGlobal } from './ssr/types'
 import type { ParsedLocation } from './location'
 import type { AnyRouteMatch } from './Matches'
 import type { NotFoundError } from './not-found'
@@ -22,6 +24,109 @@ import type {
 } from './route'
 import type { AnyRedirect } from './redirect'
 import type { AnyRouter } from './router'
+
+type RouteComponentType =
+  | 'component'
+  | 'pendingComponent'
+  | 'errorComponent'
+  | 'notFoundComponent'
+
+export function replaceRouteChunk(
+  route: AnyRoute,
+  lazyFn: AnyRoute['lazyFn'],
+): void {
+  route.lazyFn = lazyFn ?? route.lazyFn
+  route._lazy = undefined
+}
+
+function preloadComponent(
+  route: AnyRoute,
+  type: RouteComponentType,
+): Promise<void> | undefined {
+  return (route.options[type] as any)?.preload?.()
+}
+
+function loadComponents(route: AnyRoute): Promise<void> | undefined {
+  const component = preloadComponent(route, 'component')
+  const pending = preloadComponent(route, 'pendingComponent')
+  if (component && pending) {
+    return Promise.all([component, pending]).then(() => {})
+  }
+  return component ?? pending
+}
+
+export function loadRouteChunk(
+  route: AnyRoute,
+  // `false` waits only for lazy route options, before a boundary is selected.
+  componentType?: 'errorComponent' | 'notFoundComponent' | false,
+): Promise<void> | undefined {
+  const afterLazy = () =>
+    componentType === false
+      ? undefined
+      : componentType
+        ? preloadComponent(route, componentType)
+        : loadComponents(route)
+  const current = route._lazy
+  if (current) {
+    return current === true ? afterLazy() : current.then(afterLazy)
+  }
+  if (!route.lazyFn) {
+    return afterLazy()
+  }
+
+  const promise = route.lazyFn().then(
+    (lazyRoute) => {
+      // HMR clears the owner before an obsolete import can settle.
+      if (process.env.NODE_ENV === 'production' || route._lazy === promise) {
+        const { id: _id, ...options } = lazyRoute.options
+        Object.assign(route.options, options)
+        route._lazy = true
+      }
+    },
+    (error) => {
+      if (process.env.NODE_ENV === 'production' || route._lazy === promise) {
+        route._lazy = undefined
+      }
+      throw error
+    },
+  )
+  route._lazy = promise
+  return promise.then(afterLazy)
+}
+
+/** Return the structural lane through the first terminal render boundary. */
+export function _getRenderedMatches(
+  matches: Array<AnyRouteMatch>,
+): Array<AnyRouteMatch> {
+  const end =
+    matches.findIndex(
+      (match) => match.status !== 'success' || match._notFound,
+    ) + 1
+  return end && end < matches.length ? matches.slice(0, end) : matches
+}
+
+/** Return the lane whose document assets belong to the current presentation. */
+export function _getAssetMatches(
+  matches: Array<AnyRouteMatch>,
+): Array<AnyRouteMatch> {
+  let end = matches.length
+  for (let index = 0; index < end; index++) {
+    const match = matches[index]!
+    // `_assetEnd` is only ever set on hydration presentation clones that are
+    // `status: 'pending'`, `ssr: 'data-only'`, error-free, and not not-found
+    // (see hydrate.ts), and commits clear it — so its presence alone is the guard.
+    if (match._assetEnd !== undefined) {
+      end = Math.min(end, Math.max(index + 1, match._assetEnd))
+      continue
+    }
+    if (match.status !== 'success' || match._notFound) {
+      end = index + 1
+      break
+    }
+  }
+  // `end` only ever shrinks to `index + 1 >= 1`, so no zero guard is needed.
+  return end < matches.length ? matches.slice(0, end) : matches
+}
 
 declare const lanePhase: unique symbol
 
@@ -1355,12 +1460,20 @@ function commitMatches(
       match._assetEnd = undefined
     }
   }
-  const rendered = _getRenderedMatches(matches)
+  const cut = _getRenderedMatches(matches).length
   const cached = new Map<string, AnyRouteMatch>()
   const now = Date.now()
   for (const match of [...previous, ...previousCached.values()]) {
+    // Rendered-prefix ids and settled successes anywhere in the lane are
+    // authoritative: retaining an older same-id generation would shadow them
+    // at the next planning pass. Unsettled beyond-boundary matches are not —
+    // they must not evict a newer same-id preload.
     if (
-      rendered.some((candidate) => candidate.id === match.id) ||
+      matches.some(
+        (candidate, index) =>
+          candidate.id === match.id &&
+          (index < cut || candidate.status === 'success'),
+      ) ||
       match.status !== 'success'
     ) {
       continue
@@ -1825,7 +1938,7 @@ async function runClientTransaction(
 
 export async function loadClientRoute(
   router: CoordinatorRouter,
-  opts?: { sync?: boolean },
+  opts?: { sync?: boolean; _dedupe?: boolean },
 ): Promise<void> {
   let rematerialize = false
   if (process.env.NODE_ENV !== 'production') {
@@ -1846,6 +1959,20 @@ export async function loadClientRoute(
     pendingLocation?.href === location.href
       ? (pendingLocation._redirects ?? 0)
       : 0
+  // A same-location navigation joins the transaction already loading it
+  // instead of restarting its work. Reload requests never carry the flag,
+  // and same-location redirects must restart the lane they came from.
+  if (
+    opts?._dedupe &&
+    !redirects &&
+    previousOwner &&
+    !rematerialize &&
+    previousOwner[2].href === location.href &&
+    router.stores.status.get() === 'pending'
+  ) {
+    await awaitCurrent(router)
+    return
+  }
   const handoff = router._handoff
   const hydrationController = rematerialize ? undefined : handoff?.[0]()
   const preflight = new AbortController()
@@ -2156,4 +2283,384 @@ export async function preloadClientRoute(
     }
     return
   }
+}
+
+// --- SSR hydration (client entry via @tanstack/router-core/ssr/client) ---
+
+declare global {
+  interface Window {
+    [GLOBAL_TSR]?: TsrSsrGlobal
+    [GLOBAL_SEROVAL]?: any
+  }
+}
+
+export async function hydrate(router: AnyRouter): Promise<void> {
+  if (process.env.NODE_ENV !== 'production' && !window.$_TSR) {
+    throw new Error(
+      'Invariant failed: Expected to find bootstrap data on window.$_TSR, but we did not. Please file an issue!',
+    )
+  }
+  const tsr = window.$_TSR!
+
+  const adapters = router.options.serializationAdapters as
+    | Array<AnySerializationAdapter>
+    | undefined
+  if (adapters?.length) {
+    tsr.t = new Map(
+      adapters.map((adapter) => [adapter.key, adapter.fromSerializable]),
+    )
+    tsr.buffer.forEach((script) => script())
+  }
+  tsr.initialized = true
+
+  const dehydratedRouter = tsr.router
+  if (process.env.NODE_ENV !== 'production' && !dehydratedRouter) {
+    throw new Error(
+      'Invariant failed: Expected to find a dehydrated data on window.$_TSR.router, but we did not. Please file an issue!',
+    )
+  }
+  router.ssr = { manifest: dehydratedRouter!.manifest }
+  const nonce = (
+    document.querySelector('meta[property="csp-nonce"]') as
+      | HTMLMetaElement
+      | undefined
+  )?.content
+  router.options.ssr = { nonce }
+
+  const dehydratedMatches = dehydratedRouter!.matches
+
+  const controller = new AbortController()
+  const previousPreflight = router._preflight
+  router._preflight = controller
+  previousPreflight?.abort()
+  const retire = (cause?: unknown) => {
+    if (router._preflight === controller) {
+      router._preflight = undefined
+    }
+    controller.abort(cause)
+    return false
+  }
+  const isCurrent = () =>
+    (!router._tx &&
+      router._preflight === controller &&
+      !controller.signal.aborted) ||
+    retire()
+
+  let location!: AnyRouter['latestLocation']
+  let candidates!: Array<AnyRouteMatch>
+  let handoffInputs!: ReturnType<typeof laneInputs>
+  try {
+    await waitFor(
+      router.options.hydrate?.(dehydratedRouter!.dehydratedData),
+      controller.signal,
+    )
+    if (!isCurrent()) {
+      return
+    }
+    router.updateLatestLocation()
+    location = router.latestLocation
+    router.stores.location.set(location)
+    handoffInputs = laneInputs(router, location)
+    candidates = router.matchRoutes(location, {
+      _controller: controller,
+    })
+  } catch (cause) {
+    retire(cause)
+    if (cause !== controller.signal) {
+      throw cause
+    }
+  }
+  if (!isCurrent()) {
+    return
+  }
+  const committed: Array<AnyRouteMatch> = []
+  let pendingBoundary: number | undefined
+  let verifiedAssetEnd = 0
+  const retryFrom = (index: number) => {
+    // The failing route's identity is still verified, but no descendant is.
+    verifiedAssetEnd = Math.min(verifiedAssetEnd, index + 1)
+    const removed = committed.splice(index)
+    for (const match of removed) {
+      if (
+        getRoute(router, match).options.loader &&
+        (match.status === 'success' ||
+          (!match.invalid && 'loaderData' in match))
+      ) {
+        cacheLoaderMatch(
+          router,
+          {
+            ...match,
+            status: 'success',
+            error: undefined,
+            preload: true,
+          },
+          router._cache.get(match.id),
+        )
+      }
+    }
+    transferMatchResources(router, removed)
+  }
+
+  // A longer server lane is valid only when the local match already caps the
+  // branch at a global not-found boundary. Otherwise no transported work is
+  // safe to attach to the shorter client lane.
+  const shared =
+    dehydratedMatches.length > candidates.length
+      ? candidates.findIndex((match) => match._notFound) + 1
+      : dehydratedMatches.length
+  let isTerminal = false
+  for (let index = 0; index < shared; index++) {
+    const candidate = candidates[index]!
+    const dehydrated = dehydratedMatches[index]!
+    if (
+      typeof dehydrated.i !== 'string' ||
+      hydrateSsrMatchId(dehydrated.i) !== candidate.id
+    ) {
+      pendingBoundary ??= index
+      break
+    }
+    verifiedAssetEnd = index + 1
+    const route = getRoute(router, candidate)
+    if (
+      'l' in dehydrated ||
+      (dehydrated.s === 'success' &&
+        dehydrated.e === undefined &&
+        route.options.loader)
+    ) {
+      candidate.loaderData = dehydrated.l
+    }
+    candidate.status = dehydrated.s
+    candidate.ssr = dehydrated.ssr
+    route.options.ssr = candidate.ssr
+    candidate.updatedAt = dehydrated.u
+    candidate.error = dehydrated.e
+    candidate._notFound ||= dehydrated.g
+    const terminal =
+      candidate.status === 'error' ||
+      candidate.status === 'notFound' ||
+      candidate._notFound
+    if (terminal) {
+      isTerminal = true
+      committed.push(candidate)
+      if (candidate.ssr === false || candidate.ssr === 'data-only') {
+        pendingBoundary ??= index
+      }
+      break
+    }
+    if (candidate.status === 'pending') {
+      pendingBoundary ??= index
+      break
+    }
+
+    committed.push(candidate)
+    if (candidate.ssr === 'data-only') {
+      pendingBoundary ??= index
+    }
+  }
+  let verifiedContextEnd = verifiedAssetEnd
+
+  if (
+    !isTerminal &&
+    committed.length === shared &&
+    shared < candidates.length
+  ) {
+    pendingBoundary = shared
+  }
+
+  // Hooks observe structural membership. Execution remains limited to
+  // `committed`, the accepted server prefix.
+  const chunks = committed.map(async (match) => {
+    try {
+      const route = getRoute(router, match)
+      if (match._notFound) {
+        await Promise.all([
+          loadRouteChunk(route),
+          loadRouteChunk(route, 'notFoundComponent'),
+        ])
+      } else {
+        await loadRouteChunk(
+          route,
+          match.status === 'error'
+            ? 'errorComponent'
+            : match.status === 'notFound'
+              ? 'notFoundComponent'
+              : undefined,
+        )
+      }
+      return true
+    } catch {
+      return false
+    }
+  })
+  let chunkFailure = 0
+  try {
+    while (
+      chunkFailure < chunks.length &&
+      (await waitFor(chunks[chunkFailure]!, controller.signal))
+    ) {
+      chunkFailure++
+    }
+  } catch {
+    isCurrent()
+    return
+  }
+  if (!isCurrent()) {
+    return
+  }
+  if (chunkFailure < committed.length) {
+    verifiedContextEnd = Math.min(verifiedContextEnd, chunkFailure)
+    retryFrom(chunkFailure)
+  }
+
+  // The first pending match is already visible, so prepare its route context
+  // without granting its beforeLoad or loader any hydration authority.
+  const contextEnd = Math.max(
+    pendingBoundary === committed.length
+      ? committed.length + 1
+      : committed.length,
+    verifiedContextEnd,
+  )
+  for (let index = 0; index < contextEnd; index++) {
+    const match = candidates[index]!
+    const route = getRoute(router, match)
+    const parentContext =
+      candidates[index - 1]?.context ?? router.options.context ?? {}
+    let routeContext
+    if (route.options.context) {
+      try {
+        routeContext = match._ctx =
+          route.options.context({
+            deps: match.loaderDeps,
+            params: match.params,
+            context: parentContext,
+            location,
+            navigate: navigateFrom(router, location),
+            buildLocation: router.buildLocation,
+            cause: match.cause,
+            abortController: controller,
+            preload: false,
+            matches: candidates,
+            routeId: route.id,
+          }) || {}
+      } catch {
+        if (!isCurrent()) {
+          return
+        }
+        if (
+          match.status !== 'error' &&
+          match.status !== 'notFound' &&
+          !match._notFound
+        ) {
+          retryFrom(index)
+          break
+        }
+      }
+      if (!isCurrent()) {
+        return
+      }
+    }
+    match.context = {
+      ...parentContext,
+      ...routeContext,
+      ...(committed[index] && dehydratedMatches[index]!.b),
+    }
+  }
+
+  await projectLane(
+    router,
+    [location, candidates] as any,
+    controller.signal,
+    0,
+    verifiedAssetEnd,
+  )
+  if (!isCurrent()) {
+    return
+  }
+  const needsClientLoad =
+    pendingBoundary !== undefined || committed.length < shared
+  const committedMatches =
+    isTerminal && committed.length === shared ? candidates : committed
+  let presented = needsClientLoad ? candidates : committedMatches
+  let dataOnlyAssetEnd: number | undefined
+  if (needsClientLoad && pendingBoundary !== undefined) {
+    const boundary = presented[pendingBoundary]!
+    dataOnlyAssetEnd =
+      boundary.status === 'success' &&
+      boundary.ssr === 'data-only' &&
+      boundary.error === undefined &&
+      !boundary._notFound &&
+      verifiedAssetEnd > pendingBoundary + 1
+        ? verifiedAssetEnd
+        : undefined
+    presented = presented.slice()
+    presented[pendingBoundary] = {
+      ...boundary,
+      status: 'pending',
+      ssr: boundary.ssr === 'data-only' ? 'data-only' : false,
+      _assetEnd: dataOnlyAssetEnd,
+    }
+  }
+
+  const claim = () =>
+    needsClientLoad &&
+    !router._tx &&
+    router.latestLocation.state === location.state &&
+    deepEqual(handoffInputs, laneInputs(router, router.latestLocation)) &&
+    router._committed === committedMatches &&
+    committedMatches.length &&
+    !controller.signal.aborted
+      ? controller
+      : undefined
+  const handoff: NonNullable<AnyRouter['_handoff']> = [
+    claim,
+    (matches) => {
+      if (router._handoff !== handoff) {
+        return
+      }
+      const prefix = committedMatches.length
+      if (
+        !matches ||
+        !claim() ||
+        committedMatches.some((match, index) => match.id !== matches[index]?.id)
+      ) {
+        router._handoff = undefined
+        controller.abort()
+        return
+      }
+      let handoffAssetEnd = dataOnlyAssetEnd
+      if (handoffAssetEnd !== undefined) {
+        for (let index = prefix; index < handoffAssetEnd; index++) {
+          if (candidates[index]?.id !== matches[index]?.id) {
+            handoffAssetEnd =
+              index > (pendingBoundary ?? -1) + 1 ? index : undefined
+            break
+          }
+        }
+      }
+      const clones = committedMatches.map((match) => ({ ...match }))
+      if (handoffAssetEnd !== undefined) {
+        clones[pendingBoundary!]!._assetEnd = handoffAssetEnd
+      }
+      transferMatchResources(router, matches.splice(0, prefix, ...clones))
+      for (let index = prefix; index < matches.length; index++) {
+        const match = matches[index]!
+        const hydrated = candidates[index]
+        if (hydrated?.id === match.id && hydrated._ctx) {
+          match._ctx = hydrated._ctx
+        }
+        match.abortController = controller
+      }
+      return prefix
+    },
+  ]
+  router._committed = committedMatches
+  router._handoff = handoff
+  router._preflight = undefined
+  router.batch(() => {
+    router.stores.setMatches(presented)
+    router.stores.status.set('idle')
+    if (!needsClientLoad) {
+      router.stores.resolvedLocation.set(router.stores.location.get())
+    }
+  })
 }
