+++ packages/router-core/src/ssr/ssr-client.ts
@@ -1,1 +1,1 @@
-export { hydrate } from '../hydrate'
+export { hydrate } from '../load-client'
